move all fixtures to setup_fixtures
