+++ tests/tests_toolkit/network_utils.py
@@ -1,4 +1,8 @@
+import re
+from pathlib import Path
+
 from ingenialink import CanBaudrate, CanDevice
+from packaging import version
 
 from ingeniamotion import MotionController
 from tests.tests_toolkit.setups import (
@@ -46,18 +50,39 @@ def connect_to_servo_with_protocol(mc: MotionController, descriptor: SetupDescri
         raise NotImplementedError
 
 
-def load_firmware_with_protocol(mc: MotionController, descriptor: SetupDescriptor):
+def _read_fw_version(alias: str, motion_controller: MotionController):
+    _, _, fw_version, _ = motion_controller.configuration.get_drive_info_coco_moco(servo=alias)
+    return fw_version
+
+
+def is_fw_already_uploaded(alias: str, mc: MotionController, firmware_file: Path) -> bool:
+    current_fw_version = _read_fw_version(alias=alias, motion_controller=mc)[0]
+    match = re.search(r"_(\d+\.\d+\.\d+)", firmware_file.stem)
+    if match is None:
+        return False
+    file_fw_version = match.group(1)
+    try:
+        return version.parse(file_fw_version) == version.parse(current_fw_version)
+    except version.InvalidVersion:
+        return False
+
+
+def load_firmware_with_protocol(mc: MotionController, descriptor: SetupDescriptor, alias: str):
+    if is_fw_already_uploaded(alias=alias, mc=mc, firmware_file=descriptor.fw_file):
+        return
     if isinstance(descriptor, DriveEcatSetup):
         mc.communication.load_firmware_ecat(
             ifname=descriptor.ifname,
-            fw_file=descriptor.fw_file,
+            fw_file=descriptor.fw_file.as_posix(),
             slave=descriptor.slave,
             boot_in_app=descriptor.boot_in_app,
         )
     elif isinstance(descriptor, DriveCanOpenSetup):
-        mc.communication.load_firmware_canopen(fw_file=descriptor.fw_file)
+        mc.communication.load_firmware_canopen(fw_file=descriptor.fw_file.as_posix(), servo=alias)
     elif isinstance(descriptor, DriveEthernetSetup):
-        mc.communication.load_firmware_ethernet(ip=descriptor.ip, fw_file=descriptor.fw_file)
+        mc.communication.load_firmware_ethernet(
+            ip=descriptor.ip, fw_file=descriptor.fw_file.as_posix()
+        )
     else:
         raise NotImplementedError(
             f"Firmware loading not implemented for descriptor {type(descriptor)}"
